change to float

+++ notUsed/Weapon00.java
@@ -18,9 +18,9 @@ public class Weapon00 extends NormalWeapon{
             if (input==null) return;
             Vec2d pos=owner.getBulletPosition();
             Vec2d velocity=input.limit(speed*3);
-            double size=this.size*2;
-            double health=this.health*20;
-            double damage=this.damage*150;
+            float size=this.size*2;
+            float health=this.health*20;
+            float damage=this.damage*150;
             shootBullet(pos,velocity,size,health,damage);
             cooldown= (reload*0.2);
         }
