+++ src/main/java/big/engine/math/util/LineIntersection.java
@@ -7,14 +7,14 @@ public class LineIntersection {
         return isIntersect(p1.x, p1.y, p2.x, p2.y, p3.x, p3.y, p4.x, p4.y);
     }
     public static boolean isIntersect(
-            double x1, double y1, double x2, double y2,
-            double x3, double y3, double x4, double y4) {
+            float x1, float y1, float x2, float y2,
+            float x3, float y3, float x4, float y4) {
 
 
-        double d1 = direction(x3, y3, x4, y4, x1, y1);
-        double d2 = direction(x3, y3, x4, y4, x2, y2);
-        double d3 = direction(x1, y1, x2, y2, x3, y3);
-        double d4 = direction(x1, y1, x2, y2, x4, y4);
+        float d1 = direction(x3, y3, x4, y4, x1, y1);
+        float d2 = direction(x3, y3, x4, y4, x2, y2);
+        float d3 = direction(x1, y1, x2, y2, x3, y3);
+        float d4 = direction(x1, y1, x2, y2, x4, y4);
 
 
         if (((d1 > 0 && d2 < 0) || (d1 < 0 && d2 > 0)) &&
@@ -32,12 +32,12 @@ public class LineIntersection {
     }
 
 
-    private static double direction(double xi, double yi, double xj, double yj, double xk, double yk) {
+    private static float direction(float xi, float yi, float xj, float yj, float xk, float yk) {
         return (xk - xi) * (yj - yi) - (xj - xi) * (yk - yi);
     }
 
 
-    private static boolean onSegment(double xi, double yi, double xj, double yj, double xk, double yk) {
+    private static boolean onSegment(float xi, float yi, float xj, float yj, float xk, float yk) {
         return Math.min(xi, xj) <= xk && xk <= Math.max(xi, xj) &&
                Math.min(yi, yj) <= yk && yk <= Math.max(yi, yj);
     }
