+++ src/main/java/big/modules/entity/Controllable.java
@@ -5,11 +5,11 @@ import big.modules.ctrl.ServerInputManager;
 import big.modules.weapon.GunList;
 
 public interface Controllable {
-    public void setRotation(double rotation);
+    public void setRotation(float rotation);
     public ServerInputManager getInputManager();
     public Vec2d getPosition();
     public GunList getWeapon();
     Vec2d getRealVelocity();
-    double getSpeed();
-    double getFov();
+    float getSpeed();
+    float getFov();
 }
